// src/app/models/vendor-details/business-and-booth-info-table.model.ts
export interface BusinessAndBoothInfoTable {
  texasSalesTaxPermitId: string;
  inventoryChanged: boolean;
  inventoryChangedComments: string;
  remainInSameBooth: boolean;
  directMarketing: boolean;
  foodSamples: boolean;
  tables: number;
  chairs: number;
  premiumLobby: boolean;
  willingToAcceptOutsideBooth: boolean;
  willingToAcceptLastMinuteBooth: boolean;
  comments: string;
  interestedInSponsoring: number;
}